Stop sourcemapping function names
Sourcemaps will only give you the name mapping for
a specific source position. However, stack traces will
point to the actual callsite not the function declaration.
So we always ask for the name mapping of the variable
not the calling function.

We can fix this in v8 by using `getEnclosingColumnNumber() - 1`
but this is more involed.

Since all the sourcemapping currently targets dev,
we just stop sourcemapping assuming most function names
are not mangled. That holds at least true for all
product code.